\name{Fancyaxis-package}
\alias{Fancyaxis-package}
\alias{Fancyaxis}
\docType{package}
\title{
Functions for drawing scatter-plots and axes inspired by E. Tufte's redesigns.
}
\description{
A set of functions for plotting scatterplots and creating 
  axes which are inspired by E. Tufte's ""The Visual Display of 
  Quantitative Information"
}
\details{
\tabular{ll}{
Package: \tab Fancyaxis\cr
Version: \tab 0.1\cr
Date: \tab 2012-06-10\cr
License: \tab GPL version 3 or newer\cr
Built: \tab R 2.15.0; ; 2012-06-10 16:59:11 UTC; unix\cr
}
Functions in this package add features to scatterplots or histograms. 
One of the primary functions is \code{\link{fancyaxis}} which shows
statistics concerning the marginal distribution of a variable.
}
\author{
Steven J. Murdoch <http://www.cl.cam.ac.uk/users/sjm217/>

Maintainer: Peter N. Steinmetz <PeterNSteinmetz@steinmetz.org>
}
\references{
"The Visual Display of Quantitative Information", 2nd Ed., E. Tufte, 2001. 
}
\keyword{ package }
